# Conflicts:
#	cmake/Modules/Packages/ML-PACE.cmake
#	src/ML-PACE/pair_grace.cpp
#	src/ML-PACE/pair_grace.h
#	src/ML-PACE/pair_grace_fs.cpp
#	src/ML-PACE/pair_pace.cpp
#	src/ML-PACE/pair_pace_extrapolation.cpp
#	src/ML-PACE/utils_pace.h

+++ src/bond.h
@@ -48,7 +48,7 @@ class Bond : protected Pointers {
   // KOKKOS host/device flag and data masks
 
   ExecutionSpace execution_space;
-  unsigned int datamask_read, datamask_modify;
+  uint64_t datamask_read, datamask_modify;
   int copymode, kokkosable;
 
   Bond(class LAMMPS *);
@@ -61,8 +61,8 @@ class Bond : protected Pointers {
   virtual double equilibrium_distance(int) = 0;
   virtual void write_restart(FILE *) = 0;
   virtual void read_restart(FILE *) = 0;
-  virtual void write_restart_settings(FILE *){};
-  virtual void read_restart_settings(FILE *){};
+  virtual void write_restart_settings(FILE *) {};
+  virtual void read_restart_settings(FILE *) {};
   virtual void write_data(FILE *) {}
   virtual double single(int, double, int, int, double &) = 0;
   virtual double memory_usage();
